@@ -1,11 +1,12 @@
 <?php
 
 use Illuminate\Support\Facades\Route;
+use LaravelEnso\Core\Http\Controllers\Spa;
 
 Route::prefix('core')
     ->as('core.')
     ->group(function () {
-        Route::get('home', 'Spa')->name('home.index');
+        Route::get('home', Spa::class)->name('home.index');
 
-        require 'core/preferences.php';
+        require __DIR__.'/core/preferences.php';
     });
